Move cogs to their own sub-package.
