+++ roll35/cogs/magicitem.py
@@ -9,11 +9,11 @@ import random
 
 from nextcord.ext import commands
 
-from .common import ret_async, bad_return
-from .data.constants import RANK
-from .parser import Parser
-from .retcode import Ret
-from .types.cog import R35Cog
+from ..common import ret_async, bad_return
+from ..data.constants import RANK
+from ..parser import Parser
+from ..retcode import Ret
+from ..types import R35Cog
 
 NOT_READY = 'Magic item data is not yet available, please try again later.'
 NO_ITEMS_IN_COST_RANGE = 'No items found in requested cost range.'
